+++ packages/widget/src/types/widget.ts
@@ -2,6 +2,7 @@ import type {
   BaseToken,
   ChainType,
   ContractCall,
+  ExtendedChain,
   Order,
   RouteExtended,
   RouteOptions,
@@ -101,6 +102,7 @@ export enum HiddenUI {
   IntegratorStepDetails = 'integratorStepDetails',
   ReverseTokensButton = 'reverseTokensButton',
   RouteTokenDescription = 'routeTokenDescription',
+  ChainSelect = 'chainSelect',
 }
 export type HiddenUIType = `${HiddenUI}`
 
@@ -143,10 +145,10 @@ export interface WidgetContractTool {
 }
 
 export interface CalculateFeeParams {
-  fromChainId: number
-  toChainId: number
-  fromTokenAddress: string
-  toTokenAddress: string
+  fromChain: ExtendedChain
+  toChain: ExtendedChain
+  fromToken: Token
+  toToken: Token
   fromAddress?: string
   toAddress?: string
   fromAmount?: bigint
@@ -158,6 +160,20 @@ export interface WidgetFeeConfig {
   name?: string
   logoURI?: string
   fee?: number
+  /**
+   * Whether to show the fee percentage in the fee details.
+   * @default false
+   */
+  showFeePercentage?: boolean
+  /**
+   * Whether to show a tooltip with the fee details. Requires `name` or `feeTooltipComponent` to be set.
+   * @default false
+   */
+  showFeeTooltip?: boolean
+  /**
+   * Custom tooltip component to show with the fee details.
+   */
+  feeTooltipComponent?: ReactNode
   /**
    * Function to calculate fees before fetching quotes.
    * If provided, this function will be used instead of the `fee` parameter.